Player Win counts are preserved across multiple games.  Still must look into powerup audio not destroying....
Wins are carried across games and reset if the game ever returns to the
main menu. Therefore, it players play multiple games, wins are carried.
If they return to the main menu at any point, the wins are reset. Done
using PlayerPerf keys (technically its a save file but it works well)